+++ src/lagransala/extractor/domain/__init__.py
@@ -1,9 +1,18 @@
 from .event_data import EventData
-from .event_extractor import EmptyReason, EventExtractionResult, EventExtractor
+from .event_extractor import (
+    EmptyReason,
+    EventExtraction,
+    EventExtractor,
+    SourcedEventExtraction,
+)
+from .sourced_content import ContentFormat, SourcedContent
 
 __all__ = [
-    "EventData",
-    "EventExtractionResult",
-    "EventExtractor",
+    "ContentFormat",
     "EmptyReason",
+    "EventData",
+    "EventExtraction",
+    "EventExtractor",
+    "SourcedContent",
+    "SourcedEventExtraction",
 ]