+++ src/lagransala/shared/domain/__init__.py
@@ -1,5 +1,15 @@
 from .caching import CacheBackend
 from .coroutine_with_data import coroutine_with_data
 from .fetcher import Fetcher
+from .fetcher import Response as FetcherResponse
+from .flatten import flatten
+from .pydantic_types import Slug
 
-__all__ = ["CacheBackend", "coroutine_with_data", "Fetcher"]
+__all__ = [
+    "CacheBackend",
+    "Fetcher",
+    "FetcherResponse",
+    "coroutine_with_data",
+    "flatten",
+    "Slug",
+]