+++ src/lagransala/shared/infrastructure/aiohttp_fetcher.py
@@ -1,13 +1,14 @@
 import asyncio
-import logging
+from datetime import datetime
+from typing import Generator, Iterable
 
 import aiohttp
+from loguru import logger
+from pydantic import HttpUrl
 
 from ..application import cached
 from ..domain import CacheBackend
 from ..domain.fetcher import Response
-
-logger = logging.getLogger(__name__)
 
 
 class AiohttpFetcher:
@@ -23,7 +24,7 @@ class AiohttpFetcher:
         self._cache_backend = cache_backend
         self._cache_ttl = cache_ttl
 
-    async def fetch(self, url: str) -> Response:
+    async def fetch(self, url: HttpUrl) -> Response:
         if self._cache_backend:
             return await cached(
                 backend=self._cache_backend,
@@ -32,24 +33,25 @@ class AiohttpFetcher:
             )(self._fetch)(url=url)
         return await self._fetch(url)
 
-    async def _fetch(self, url: str) -> Response:
-        """
-        Fetches data from the given URL asynchronously.
-        :param url: The URL to fetch data from.
-        :return: The response text from the URL.
-        """
-        async with self._client.get(url) as response:
-            return Response(
-                status=response.status,
-                content=await response.text(),
-                content_type=response.content_type,
-            )
-
-    async def fetch_urls(self, urls: list[str]) -> list[Response]:
-        """
-        Fetches data from multiple URLs concurrently.
-        :param urls: List of URLs to fetch data from.
-        :return: List of responses from the URLs.
-        """
+    async def _fetch(self, url: HttpUrl) -> Response:
+        logger.debug(f"fetching {url}")
         async with self._semaphore:
-            return await asyncio.gather(*[self.fetch(url) for url in urls])
+            async with self._client.get(str(url)) as response:
+                return Response(
+                    url=url,
+                    status=response.status,
+                    content=await response.text(),
+                    content_type=response.content_type,
+                    dt=datetime.now(),
+                )
+
+    async def fetch_urls(self, urls: Iterable[HttpUrl]) -> list[Response]:
+        return await asyncio.gather(*[self.fetch(url) for url in urls])
+
+    def fetch_task(self, url: HttpUrl) -> asyncio.Task[Response]:
+        return asyncio.create_task(self.fetch(url))
+
+    def fetch_urls_tasks(
+        self, urls: Iterable[HttpUrl]
+    ) -> Generator[asyncio.Task[Response]]:
+        return (self.fetch_task(url) for url in urls)